# Normalize request path to NFC and remove/resolve dot segments (#273).