--- a/test/Microsoft.AspNetCore.Server.KestrelTests/ServerAddressFacts.cs
+++ b/test/Microsoft.AspNetCore.Server.KestrelTests/ServerAddressFacts.cs
@@ -1,3 +1,8 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Text;
 using Microsoft.AspNetCore.Server.Kestrel;
 using Xunit;
 
@@ -42,6 +47,15 @@ namespace Microsoft.AspNetCore.Server.KestrelTests
             Assert.Equal(host, serverAddress.Host);
             Assert.Equal(port, serverAddress.Port);
             Assert.Equal(pathBase, serverAddress.PathBase);
+        }
+
+        [Fact]
+        public void PathBaseIsNormalized()
+        {
+            var serverAddres = ServerAddress.FromUrl("http://localhost:8080/p\u0041\u030Athbase");
+
+            Assert.True(serverAddres.PathBase.IsNormalized(NormalizationForm.FormC));
+            Assert.Equal("/p\u00C5thbase", serverAddres.PathBase);
         }
     }
 }