@halter73 2016-01-13
I'm curious to see how this affects perf for requests that don't need normalization (e.g. the plaintext benchmark).