@benaadams 2016-01-13
It shouldn't effect it much as `needDecode` would be `false` it that case and it should just skip it?