@halter73 2016-01-13
There shouldn't be much of an affect as long as `PathNormalizer.NeedsNormalization` isn't too expensive. I would like to know for sure that's the case.