@cesarbs 2016-01-19
Comments addressed, perf test still pending.