@cesarbs 2016-01-20
There's a small perf hit in the plain text benchmark:

Before:

    Running 10s test @ http://10.0.0.100:5001/plaintext
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     4.67ms   12.42ms 187.86ms   94.60%
        Req/Sec    34.92k     3.16k   61.86k    77.44%
      11197736 requests in 10.10s, 1.38GB read
    Requests/sec: 1108681.79
    Transfer/sec:    139.57MB

    Running 10s test @ http://10.0.0.100:5001/plaintext
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     3.96ms    8.65ms 127.63ms   91.75%
        Req/Sec    34.84k     3.09k   44.80k    72.59%
      11194786 requests in 10.10s, 1.38GB read
    Requests/sec: 1108504.52
    Transfer/sec:    139.54MB

    Running 10s test @ http://10.0.0.100:5001/plaintext
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     9.66ms   22.11ms 378.68ms   89.15%
        Req/Sec    34.24k     3.55k   59.86k    73.48%
      10997567 requests in 10.10s, 1.35GB read
    Requests/sec: 1088929.91
    Transfer/sec:    137.08MB

After:

    Running 10s test @ http://10.0.0.100:5001/plaintext
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     7.54ms   22.43ms 334.46ms   93.61%
        Req/Sec    33.75k     3.49k   64.74k    78.30%
      10834047 requests in 10.10s, 1.33GB read
    Requests/sec: 1072774.90
    Transfer/sec:    135.05MB

    Running 10s test @ http://10.0.0.100:5001/plaintext
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     6.06ms   15.59ms 285.80ms   92.87%
        Req/Sec    33.99k     2.79k   51.70k    74.58%
      10911305 requests in 10.10s, 1.34GB read
    Requests/sec: 1080355.79
    Transfer/sec:    136.00MB

    Running 10s test @ http://10.0.0.100:5001/plaintext
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     4.93ms   11.70ms 254.75ms   92.00%
        Req/Sec    33.52k     3.10k   49.13k    74.06%
      10763402 requests in 10.10s, 1.32GB read
    Requests/sec: 1065676.09
    Transfer/sec:    134.15MB

Ratio between worse after and best before is 0.9786.