@cesarbs 2016-01-20
The hit is significant when decoding + normalization are needed:

Before:

    Running 10s test @ http://10.0.0.100:5001/plaintext/%41%CC%8A
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     3.79ms    7.68ms 200.81ms   96.34%
        Req/Sec    29.86k     3.22k   51.04k    73.78%
      9585223 requests in 10.10s, 1.18GB read
    Requests/sec: 949067.03
    Transfer/sec:    119.47MB

    Running 10s test @ http://10.0.0.100:5001/plaintext/%41%CC%8A
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     4.21ms    9.23ms 129.26ms   94.94%
        Req/Sec    30.78k     2.88k   46.97k    78.71%
      9882682 requests in 10.10s, 1.21GB read
    Requests/sec: 978584.03
    Transfer/sec:    123.19MB

    Running 10s test @ http://10.0.0.100:5001/plaintext/%41%CC%8A
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     4.91ms   12.42ms 235.40ms   94.12%
        Req/Sec    31.24k     2.58k   49.29k    77.63%
      10036237 requests in 10.10s, 1.23GB read
    Requests/sec: 993710.05
    Transfer/sec:    125.09MB

After:

    Running 10s test @ http://10.0.0.100:5001/plaintext/%41%CC%8A
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     3.75ms    3.77ms  94.38ms   94.26%
        Req/Sec    22.60k     1.53k   32.20k    79.50%
      7252043 requests in 10.10s, 0.89GB read
    Requests/sec: 718004.23
    Transfer/sec:     90.39MB

    Running 10s test @ http://10.0.0.100:5001/plaintext/%41%CC%8A
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     3.56ms    3.10ms  58.15ms   91.99%
        Req/Sec    22.61k     1.54k   31.35k    77.02%
      7255969 requests in 10.10s, 0.89GB read
      Socket errors: connect 0, read 0, write 81, timeout 0
    Requests/sec: 718417.96
    Transfer/sec:     90.44MB

    Running 10s test @ http://10.0.0.100:5001/plaintext/%41%CC%8A
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     3.51ms    2.89ms  54.24ms   88.69%
        Req/Sec    22.50k     1.49k   32.51k    78.47%
      7218541 requests in 10.10s, 0.89GB read
    Requests/sec: 714747.15
    Transfer/sec:     89.98MB