@benaadams 2016-01-20
When just normalization?