@cesarbs 2016-01-20
@benaadams NFC normalization will only take place in paths with percent-encoded characters.

I'm making a small change that might make a difference - I'll only try to remove dot segments if I detect any. This will save cycles and allocations.