@halter73 2016-01-20
Do you have any data for normalizing paths?