@cesarbs 2016-01-20
Got some perf back by avoiding dot segment removal when not necessaryL

    Running 10s test @ http://10.0.0.100:5001/plaintext/%41%CC%8A
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     3.91ms    5.80ms 173.92ms   95.48%
        Req/Sec    24.26k     1.88k   37.44k    75.29%
      7790526 requests in 10.10s, 0.96GB read
    Requests/sec: 771378.77
    Transfer/sec:     97.11MB

    Running 10s test @ http://10.0.0.100:5001/plaintext/%41%CC%8A
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     3.52ms    4.20ms 102.13ms   94.37%
        Req/Sec    24.24k     1.69k   36.02k    77.64%
      7780465 requests in 10.10s, 0.96GB read
    Requests/sec: 770358.86
    Transfer/sec:     96.98MB

    Running 10s test @ http://10.0.0.100:5001/plaintext/%41%CC%8A
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     3.63ms    4.17ms 113.08ms   93.35%
        Req/Sec    24.48k     1.76k   36.00k    74.83%
      7859579 requests in 10.10s, 0.97GB read
    Requests/sec: 778191.01
    Transfer/sec:     97.96MB

@Tratcher suggested a better change which I'll implement now.